Improve handling of Colby password
This allows the passing of a Colby password to the wiki page fetcher,
which eliminates the need to entire the same password twice.

It also stops the immediate prompt for password before file is even
interpreted, meaning that the password is only requested after the
arguments have been successfully parsed.