Allow manually specified path to PDF printer
Previously, I tried to add a layer of abstraction, automatically calling
wkhtmltopdf or phantomjs based off a string indicating which to use.
It's more intuitive to just let the configuration specify a full path to
the printer, and try to pick up which is which from that path.

Also fixes faulty verbosity behavior in wkhtmltopdf.